2_Multi_Agent_P2P_NATS/Client/requirements.txt
uvicorn
httpx
langgraph
langchain
langchain-openai
langchain-community
pydantic
python-dotenv
tavily-python
nats-py